+++ CHANGELOG.md
@@ -53,6 +53,7 @@
   * The method `registerSearchInput` has been renamed to `registerQuickFilter`.
   * The public attribute `searchInputComponent` has been renamed to `quickFilterComponent`.
 * **o-table**: Attribute `show-table-buttons-text` has been renamed to `show-buttons-text`.
+* **o-form**: Output `onFormDataLoaded` has been renamed to `onDataLoaded`.
 * **o-user-info**: removed input `use-flag-icons`. This component doesn't contain the language selector anymore.
 * Changes in CSS classes naming:
   * **o-app-layout**: Internal CSS classes changes: